stproject/requirements.txt
asgiref==3.7.2
autopep8==1.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclass-wizard==0.22.3
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-import-export==3.3.7
django-jazzmin==2.6.0
django-material-admin==1.8.6
django-restframework==0.0.1
django-taggit==5.0.1
django-unixdatetimefield==1.2.0
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
idna==3.6
mailhog==0.1.0
MarkupPy==1.14
myenv==201902.7
odfpy==1.4.1
openpyxl==3.1.2
pbr==6.0.0
Pillow==9.5.0
platformdirs==4.2.0
pycodestyle==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
stevedore==5.1.0
tablib==3.5.0
toml==0.10.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
xlrd==2.0.1
xlwt==1.3.0
